refactor(ver2): move archive date parsing into archiveDate module
Extract SPC archive date normalization out of sources.ts so CodeScene
delta checks are not tripped by helper duplication or argument-count
issues in the source adapter module.
